# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Neelgund Developers API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export type UserProfileRole =
  (typeof UserProfileRole)[keyof typeof UserProfileRole];

export const UserProfileRole = {
  admin: "admin",
  manager: "manager",
  employee: "employee",
} as const;

export interface UserProfile {
  id: string;
  name: string;
  email: string;
  role: UserProfileRole;
  createdAt: string;
}

export interface LoginResponse {
  token: string;
  user: UserProfile;
}

export type CreateUserRequestRole =
  (typeof CreateUserRequestRole)[keyof typeof CreateUserRequestRole];

export const CreateUserRequestRole = {
  admin: "admin",
  manager: "manager",
  employee: "employee",
} as const;

export interface CreateUserRequest {
  name: string;
  email: string;
  password: string;
  role: CreateUserRequestRole;
}

export type LeadStatus = (typeof LeadStatus)[keyof typeof LeadStatus];

export const LeadStatus = {
  new: "new",
  follow_up: "follow_up",
  meeting_scheduled: "meeting_scheduled",
  negotiation: "negotiation",
  closed_won: "closed_won",
  closed_lost: "closed_lost",
} as const;

export interface Lead {
  id: number;
  employeeId: string;
  employeeName?: string;
  name: string;
  phone: string;
  email?: string | null;
  propertyInterest?: string | null;
  status: LeadStatus;
  notes?: string | null;
  latitude?: number | null;
  longitude?: number | null;
  createdAt: string;
  updatedAt: string;
}

export type CreateLeadRequestStatus =
  (typeof CreateLeadRequestStatus)[keyof typeof CreateLeadRequestStatus];

export const CreateLeadRequestStatus = {
  new: "new",
  follow_up: "follow_up",
  meeting_scheduled: "meeting_scheduled",
  negotiation: "negotiation",
  closed_won: "closed_won",
  closed_lost: "closed_lost",
} as const;

export interface CreateLeadRequest {
  name: string;
  phone: string;
  email?: string | null;
  propertyInterest?: string | null;
  status?: CreateLeadRequestStatus;
  notes?: string | null;
  /** GPS latitude — required, captured at moment of lead entry */
  latitude: number;
  /** GPS longitude — required, captured at moment of lead entry */
  longitude: number;
}

export type UpdateLeadRequestStatus =
  (typeof UpdateLeadRequestStatus)[keyof typeof UpdateLeadRequestStatus];

export const UpdateLeadRequestStatus = {
  new: "new",
  follow_up: "follow_up",
  meeting_scheduled: "meeting_scheduled",
  negotiation: "negotiation",
  closed_won: "closed_won",
  closed_lost: "closed_lost",
} as const;

export interface UpdateLeadRequest {
  name?: string;
  phone?: string;
  email?: string | null;
  propertyInterest?: string | null;
  status?: UpdateLeadRequestStatus;
  notes?: string | null;
}

export interface LeadMeeting {
  id: number;
  leadId: number;
  scheduledAt: string;
  notes?: string | null;
  createdAt: string;
}

export interface CreateLeadMeetingRequest {
  scheduledAt: string;
  notes?: string | null;
}

export interface Document {
  id: number;
  name: string;
  url: string;
  mimeType?: string | null;
  uploadedBy: string;
  createdAt: string;
}

export interface AddDocumentRequest {
  name: string;
  url: string;
  mimeType?: string | null;
}

export interface TrackLocationRequest {
  latitude: number;
  longitude: number;
  accuracy?: number | null;
  recordedAt?: string | null;
}

export interface LocationPoint {
  id: number;
  employeeId: string;
  latitude: number;
  longitude: number;
  accuracy?: number | null;
  recordedAt: string;
}

export interface LocationTrailResponse {
  points: LocationPoint[];
  /** Array of [longitude, latitude] coordinate pairs from Mapbox Map Matching */
  matchedRoute?: number[][];
  /** Confidence score from Mapbox (0-1) */
  matchConfidence?: number | null;
  /** Whether map matching is still processing */
  matchPending: boolean;
}

export type MatchRouteRequestPointsItem = {
  lat: number;
  lng: number;
  timestamp: string;
};

export interface MatchRouteRequest {
  employeeId: string;
  /** Date in YYYY-MM-DD format */
  date: string;
  points: MatchRouteRequestPointsItem[];
}

export interface MatchRouteResponse {
  /** Array of [longitude, latitude] coordinate pairs */
  coordinates: number[][];
  /** Mapbox match confidence (0-1) */
  confidence: number;
}

export interface HeartbeatRequest {
  trackerState: string;
  platform: string;
}

export interface HeartbeatResponse {
  ok: boolean;
  serverTime: string;
  trackerState?: string;
  platform?: string;
}

export interface EmployeeLocation {
  employeeId: string;
  employeeName: string;
  latitude: number;
  longitude: number;
  recordedAt: string;
}

export interface CompanyDocument {
  id: number;
  name: string;
  url: string;
  mimeType?: string | null;
  category?: string | null;
  uploadedBy: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateCompanyDocumentRequest {
  name: string;
  url: string;
  mimeType?: string | null;
  category?: string | null;
}

export interface LeaderboardEntry {
  employeeId: string;
  employeeName: string;
  totalLeads: number;
  closedWon: number;
  rank: number;
}

export type ZoneType = (typeof ZoneType)[keyof typeof ZoneType];

export const ZoneType = {
  circle: "circle",
  polygon: "polygon",
} as const;

export interface Zone {
  id: number;
  name: string;
  type: ZoneType;
  color?: string | null;
  centerLatitude?: number | null;
  centerLongitude?: number | null;
  radiusMeters?: number | null;
  polygonGeoJson?: string | null;
  createdAt?: string | null;
}

export type CreateZoneRequestType =
  (typeof CreateZoneRequestType)[keyof typeof CreateZoneRequestType];

export const CreateZoneRequestType = {
  circle: "circle",
  polygon: "polygon",
} as const;

export interface CreateZoneRequest {
  name: string;
  type: CreateZoneRequestType;
  color?: string | null;
  centerLatitude?: number | null;
  centerLongitude?: number | null;
  radiusMeters?: number | null;
  polygonGeoJson?: string | null;
}

export type UpdateZoneRequestType =
  (typeof UpdateZoneRequestType)[keyof typeof UpdateZoneRequestType];

export const UpdateZoneRequestType = {
  circle: "circle",
  polygon: "polygon",
} as const;

export interface UpdateZoneRequest {
  name: string;
  type: UpdateZoneRequestType;
  color?: string | null;
  centerLatitude?: number | null;
  centerLongitude?: number | null;
  radiusMeters?: number | null;
  polygonGeoJson?: string | null;
}

export interface ZoneVisit {
  id: number;
  zoneId: number;
  employeeId: string;
  enteredAt: string;
  exitedAt?: string | null;
  locationPointId?: number | null;
  employeeName?: string | null;
  zoneName?: string | null;
}

export type GetLocationTrailParams = {
  employeeId: string;
  date: string;
  matchRoads?: boolean;
};

export type GetEmployeesByDateParams = {
  /**
   * Date in YYYY-MM-DD format
   */
  date: string;
};