# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Neelgund Developers API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /company-documents:
    get:
      operationId: listCompanyDocuments
      tags: [company-documents]
      summary: List company documents
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of company documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CompanyDocument"
    post:
      operationId: createCompanyDocument
      tags: [company-documents]
      summary: Upload a company document (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCompanyDocumentRequest"
      responses:
        "201":
          description: Document created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyDocument"
  /company-documents/{id}:
    put:
      operationId: updateCompanyDocument
      tags: [company-documents]
      summary: Update a company document (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCompanyDocumentRequest"
      responses:
        "200":
          description: Document updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompanyDocument"
    delete:
      operationId: deleteCompanyDocument
      tags: [company-documents]
      summary: Delete a company document (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Document deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
      required:
        - token
        - user
    UserProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, manager, employee]
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - role
        - createdAt
...
    CompanyDocument:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        url:
          type: string
        mimeType:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        uploadedBy:
          type: string
          format: uuid
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - url
        - uploadedBy
        - createdAt
        - updatedAt
    CreateCompanyDocumentRequest:
      type: object
      properties:
        name:
          type: string
        url:
          type: string
        mimeType:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
      required:
        - name
        - url
...